introduce distros.bsd.BSD class
The BSD distros will now inherite from this base BSD class.
